DB pool method changed.
